#31 refactor: 디렉터리 구조 정리 및 배럴 익스포트 적용
