+++ src/middleware.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { createServerClient } from '@supabase/ssr';
-import type { Role } from '@/shared/auth/types/role.types';
-import { envServer } from '@/shared/lib/env.server';
+import type { Role } from '@/shared/auth/types';
+import { envServer } from '@/shared/lib/env/env.server';
 
 /** 공개 경로: 루트, 로그인, /public/** */
 const PUBLIC_PATHS = [/^\/$/, /^\/login$/, /^\/public(\/|$)/];
