+++ src/shared/auth/components/index.ts
@@ -1,1 +1,3 @@
-export { default as KakaoLoginButton } from "./KakaoLoginButton";
+export { default as KakaoLoginButton } from './KakaoLoginButton';
+export * from './SignOutButton';
+export * from './UserInfo';
